Move window texture to the global frame gui part.
