+++ editor/CMakeLists.txt
@@ -7,8 +7,6 @@ add_executable(FrameEditor
     menubar.h
     menubar_view.cpp
     menubar_view.h
-    window_texture.cpp
-    window_texture.h
     ${CMAKE_CURRENT_SOURCE_DIR}/../asset/json/editor.json
 )
