+++ include/frame/gui/window_texture.h
